requirements.txt

# Async support
aiohttp>=3.8.0
aiosqlite>=0.19.0
altair>=5.0.0
anthropic>=0.7.0
asyncio-mqtt>=0.13.0
asyncio-throttle>=1.0.0
beautifulsoup4>=4.12.0

# New development dependencies
black>=23.7.0
click>=8.1.0

# New monitoring and logging dependencies
colorama>=0.4.6
discord.py>=2.3.0
emoji>=2.8.0

# Web scraping and APIs
feedparser>=6.0.10
flake8>=6.0.0
flask>=2.3.0
functions-framework>=3.4.0
google-api-core>=2.25.1
google-auth>=2.40.3

# New Google Cloud dependencies
google-cloud-functions>=1.8.0
google-cloud-scheduler>=2.8.0

# Google Cloud dependencies
google-cloud-storage>=3.1.1
google-generativeai>=0.3.0
httpx>=0.24.0
humanize>=4.8.0
joblib>=1.3.0
lightgbm>=4.0.0
lxml>=4.9.0
matplotlib>=3.7.0
matplotlib>=3.7.0
mypy>=1.5.0
nest-asyncio>=1.5.0
nltk>=3.8.1

# Notion integration
notion-client>=2.0.0
numpy>=1.24.0

# New AI and ML dependencies
openai>=1.0.0
pandas>=2.0.0

# New dependencies from the code block
Pillow>=10.0.0

# Playwright for web scraping
playwright>=1.40.0

# Visualization
plotly>=5.15.0
plotly>=5.15.0
pre-commit>=3.3.0
psutil>=5.9.0

# New data processing dependencies
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Additional dependencies
pypandoc>=1.12
PyPDF2>=3.0.0
pytest>=8.2.0

# New testing dependencies
pytest-cov>=4.1.0
pytest-httpx>=0.30.0
pytest-mock>=3.11.0

# Additional utilities
python-dateutil>=2.8.2
python-dotenv>=1.0.0
pytz>=2023.3
pyyaml>=6.0

# PDF and document processing
reportlab>=4.0.0
requests>=2.31.0
rich>=13.7.0

# Data processing and ML
scikit-learn>=1.3.0
seaborn>=0.12.0
seaborn>=0.12.0

# New web scraping dependencies
selenium>=4.15.0
sqlalchemy>=2.0.0
sqlmodel>=0.0.8
# Core dependencies
streamlit>=1.28.0
streamlit-autorefresh>=0.0.2

# Logging and monitoring
structlog>=23.1.0
telethon>=1.28.0
textblob>=0.17.1
torch>=2.0.0

# New utilities
tqdm>=4.65.0
transformers>=4.30.0
vaderSentiment==3.3.2

# New PDF generation dependencies
weasyprint>=60.0
webdriver-manager>=4.0.0

# Enhanced ML libraries
xgboost>=1.7.0
